+++ src/components/form/input/Input.stories.tsx
@@ -1,0 +1,59 @@
+import React from "react";
+
+import type { StoryObj, Meta } from "@storybook/react";
+
+import { Icon, ICON } from "../../icons/Icon.js";
+import { countries } from "../../../demo/components/form/countries.const";
+
+import { Input } from "./Input.js";
+
+const meta: Meta<typeof Input> = {
+    title: "Components/Form/Input",
+    component: Input,
+    tags: ["autodocs", "form"],
+    argTypes: {
+        prefix: {
+            type: "string",
+        },
+        suffix: {
+            type: "string",
+        },
+    },
+};
+
+type Story = StoryObj<typeof Input>;
+
+const Primary: Story = {
+    args: {
+        placeholder: "Capacity",
+        prefix: <Icon name={ICON.battery} />,
+        suffix: "kWh",
+    },
+};
+
+const handleSuggestionMatch = (s: string, picked: boolean) => {
+    alert("Suggestion " + (picked ? "picked" : "matched") + " : " + s);
+};
+
+const Mixed: Story = {
+    render: () => (
+        <>
+            <Input placeholder={"Capacity"} prefix={<Icon name={ICON.battery} />} suffix={"kWh"} />
+            <Input value={"William J. Welter"} />
+            <Input prefix={"Text prefix"} value={"Polly W. Wilson"} />
+            <Input prefix={"Disabled"} value={"Clayton J. Foster"} disabled={true} />
+            <Input prefix={"Read only"} value={"Norma J. Gomez"} readOnly={true} />
+            <Input
+                placeholder={"With suggestions, try country name"}
+                suggestions={countries}
+                onSuggestionMatch={handleSuggestionMatch}
+            />
+        </>
+    ),
+};
+
+export {
+    Primary, Mixed,
+};
+
+export default meta;