+++ src/index.ts
@@ -26,6 +26,11 @@ export * from "./components/ui/button/Button.js";
 export * from "./components/ui/directionPad/Pad.js";
 export * from "./components/ui/drawer/Drawer.js";
 export * from "./components/ui/keyValue/KeyValue.js";
+export * from "./components/ui/loader/CoveringLoader.js";
+export * from "./components/ui/loader/FullLoader.js";
+export * from "./components/ui/loader/Loader.js";
+export * from "./components/ui/loader/Loading.js";
+export * from "./components/ui/loader/PopLoader.js";
 export * from "./components/ui/message/Message.js";
 export * from "./components/ui/modal/Modal.js";
 export * from "./components/ui/modal/ModalButtons.js";
@@ -40,4 +45,7 @@ export * from "./components/utils/HandleEsc.js";
 export * from "./components/utils/If.js";
 export * from "./components/utils/Spacer.js";
 
+export type { ThemeCSS } from "./theme.js";
+
 export { getCssText, styled } from "./theme.js";
+export { cssReset } from "./theme.css-reset.js";